docs(stepper): update stepper demo
re #413
