+++ src/stepper/__test__/demo.test.jsx
@@ -6,15 +6,19 @@ import { mount } from '@vue/test-utils';
 import baseVue from '@/stepper/demos/base.vue';
 import disableInputVue from '@/stepper/demos/disable-input.vue';
 import disabledVue from '@/stepper/demos/disabled.vue';
-import eventVue from '@/stepper/demos/event.vue';
 import mobileVue from '@/stepper/demos/mobile.vue';
+import pureStepperVue from '@/stepper/demos/pure-stepper.vue';
+import statusVue from '@/stepper/demos/status.vue';
+import unitStepperVue from '@/stepper/demos/unit-stepper.vue';
 
 const mapper = {
   baseVue,
   disableInputVue,
   disabledVue,
-  eventVue,
   mobileVue,
+  pureStepperVue,
+  statusVue,
+  unitStepperVue,
 };
 
 describe('Stepper', () => {
