--- src/stepper/demos/disabled.vue
@@ -1,13 +1,0 @@
-<template>
-  <div class="stepper-demo">
-    <t-cell title="禁用">
-      <t-stepper v-model="number" :step="1" :min="0" :max="100" disabled />
-    </t-cell>
-  </div>
-</template>
-
-<script lang="ts" setup>
-import { ref } from 'vue';
-
-const number = ref(0);
-</script>
